+++ slack-gamebot/commands/season.rb
@@ -2,8 +2,10 @@ module SlackGamebot
   module Commands
     class Season < SlackRubyBot::Commands::Base
       def self.call(client, data, _match)
-        current_season = ::Season.new
-        if current_season.valid?
+        if ::Season.count > 0 && !::Challenge.current.any?
+          send_message_with_gif client, data.channel, 'No matches have been recorded.', 'history'
+        elsif ::Challenge.current.any?
+          current_season = ::Season.new
           send_message client, data.channel, current_season.to_s
         else
           send_message_with_gif client, data.channel, "There're no seasons.", %w(winter summer fall spring).sample